Stedman
Stedman on odd stages now has a simpler syntax, as shown in test_stedman() in test_prover.py

New built in post_proof for generating rows after composition truth has been determined. Designed for use with bracketted courseheads it is executed on a new composition object. "true" is executed afterwards, but on the original composition.