External Rows
New syntax (! <arguments> !) for executing rows outside the current proof, e.g. for generating bracketed courseends. Any place notation inside the brackets is processed on a new composition object starting from the current row of the proof. The length of this composition is initially 0.
post_proof when called at the end of a proof operates similarly to the nex syntax, though the (! !) is not required around it at present.